+++ DuneEngine/src/Dune/Core/Input.h
@@ -224,11 +224,11 @@ namespace Dune
 		inline static bool GetMouseButtonDown(int index) { return m_mouseDown[index]; }
 		inline static bool GetMouseButtonUp(int index) { return m_mouseUp[index]; }
 
-		inline static double GetMousePosX() { return m_mousePosX; }
-		inline static double GetMousePosY() { return m_mousePosY; }
-		inline static double GetMouseDeltaX() { return m_mouseDeltaX; }
-		inline static double GetMouseDeltaY() { return m_mouseDeltaY; }
-		inline static double GetMouseWheelDelta() { return m_mouseWheelDelta; }
+		inline static float GetMousePosX() { return m_mousePosX; }
+		inline static float GetMousePosY() { return m_mousePosY; }
+		inline static float GetMouseDeltaX() { return m_mouseDeltaX; }
+		inline static float GetMouseDeltaY() { return m_mouseDeltaY; }
+		inline static float GetMouseWheelDelta() { return m_mouseWheelDelta; }
 
 		//TODO: shouldn't be able to SetKey or SetMouse outside from window event, friend class Window ?
 		static void SetKeyDown(KeyCode key);
@@ -237,9 +237,9 @@ namespace Dune
 		static void SetMouseButtonDown(int index);
 		static void SetMouseButtonUp(int index);
 
-		static void SetMousePosX(double posX);
-		static void SetMousePosY(double posY);
-		static void SetMouseWheelDelta(double delta);
+		static void SetMousePosX(float posX);
+		static void SetMousePosY(float posY);
+		static void SetMouseWheelDelta(float delta);
 		
 		static void EndFrame();
 
@@ -253,11 +253,11 @@ namespace Dune
 		static bool m_mouseDown[3];
 		static bool m_mouseUp[3];
 
-		static double m_mousePosX;
-		static double m_mousePosY;
-		static double m_mouseDeltaX;
-		static double m_mouseDeltaY;
-		static double m_mouseWheelDelta;
+		static float m_mousePosX;
+		static float m_mousePosY;
+		static float m_mouseDeltaX;
+		static float m_mouseDeltaY;
+		static float m_mouseWheelDelta;
 	};
 }
